VP-1218: split file to separated classes
DefaultableDictionaryExtensions, DefaultableDictionary
